refactor : refactorize code for webpack
