+++ interface/banned/namu_live.html
@@ -45,6 +45,5 @@
     </main>
 </body>
 <script src="/polyfill/browser-polyfill.min.js"></script>
-<script src="/lib/util.js"></script>
 <script src="/lib/banned.js"></script>
 </html>
