+++ interface/help/adblock.html
@@ -202,7 +202,5 @@
     </footer>
 </body>
 <script src="/polyfill/browser-polyfill.min.js"></script>
-<script src="/lib/util.js"></script>
-<script src="/lib/settings_util.js"></script>
 <script src="./showVersion.js"></script>
 </html>
