+++ interface/popup/index.html
@@ -219,7 +219,5 @@
     </footer>
 </body>
 <script src="/polyfill/browser-polyfill.min.js"></script>
-<script src="/lib/util.js"></script>
-<script src="/lib/settings_util.js"></script>
 <script src="/lib/popup.js"></script>
 </html>
