+++ interface/settings/index.html
@@ -280,7 +280,5 @@
     </footer>
 </body>
 <script src="/polyfill/browser-polyfill.min.js"></script>
-<script src="/lib/util.js"></script>
-<script src="/lib/settings_util.js"></script>
 <script src="/lib/settings.js"></script>
 </html>
