+++ src/popup.ts
@@ -1,3 +1,4 @@
+import { showVersion, bgconsole, setHook, updateIntelliBan } from './util'
 
 const header = document.getElementsByTagName("header")[0];
 
@@ -39,8 +40,8 @@ showVersion();
 
 
 (async () => {
+    let config = await browser.storage.sync.get() as unknown as ConfigInterface;
     try {
-        config = await browser.storage.sync.get() as unknown as ConfigInterface;
         bgconsole.log(`로드 완료. ${JSON.stringify(config)}`);
         setHook(popup_settings as HTMLInputElement[], (config) => {
             popup_updateHeader(config);
