+++ src/settings.ts
@@ -1,4 +1,4 @@
-
+import { showVersion, setHook, updateIntelliBan } from './util'
 const full_settings = [
   document.getElementById('riss_auto'),
   document.getElementById('dbpia_auto'),
@@ -15,14 +15,14 @@ const full_settings = [
 showVersion();
 
 (async () => {
-    try {
-        config = await browser.storage.sync.get() as unknown as ConfigInterface;
-        console.log(`로드 완료. ${JSON.stringify(config)}`);
-        setHook(full_settings as HTMLInputElement[], (config) => {
-          updateIntelliBan(config, (document.getElementById('intelliBan_url') as HTMLInputElement).value);
-        });
-    } catch (e) {
-        alert(e);
-        console.error(`로드 실패. ${JSON.stringify(config)}`);
-    }
+  let config = await browser.storage.sync.get() as unknown as ConfigInterface;
+  try {
+    console.log(`로드 완료. ${JSON.stringify(config)}`);
+    setHook(full_settings as HTMLInputElement[], (config) => {
+      updateIntelliBan(config, (document.getElementById('intelliBan_url') as HTMLInputElement).value);
+    });
+  } catch (e) {
+    alert(e);
+    console.error(`로드 실패. ${JSON.stringify(config)}`);
+  }
 })();
